ABD-41: Add event handler to read messages from a queue.
Added a simple function to fetch messages from a queue using boto3, and then
print them. When there are no more messages, the function ends.

We've also added a dependency on moto, which we use in order to create a
mock SQS Queue to test against.

Finally, the run.sh file has been removed, as to run this locally would
require being setup to run against genuine AWS resources (which is quite
a lot of effort to maintain - and perhaps not a good thing)

Solo: @michaelwalker