ABD-41: Unmarshal to event, and remove messages from the queue only if we succeed.
Add an Event class, representing the row that we will write to the database.
Delete the messages from the queue only if we are able to build such an
object from the JSON string.

If any errors occur, then simply log them, and go fetch the next message.
We never want an error thrown processing a single message to bring down the
lambda. Instead, just log the error, leave it in the queue, and let SQS's
retry / dead letter queue pattern handle things for us.

Solo: @michaelwalker