ABD-41: Tests now run in a docker container with postgres running.
We want to be able to run postgres locally in order to test that we can
correctly write to a Database. (Postgres is a good match for RDS running
Postgres)

Ideally, we would do this by spinning up a side-car docker container running
just postgres, and connect to that, but unfortunately side-car containers
are not secure in jenkins. We are therefore running postgres directly
within our python test-runner container.

Pre-commit now builds the container for us, and runs tests from within the
container, rather than running them directly on your local system.

Things that aren't yet ideal.
- it prints the output 'event-emitter-container' quite a lot
- it re-installs requirements.txt each time, which is a little slow.

Solo: @michaelwalker.