ABD-41: Don't throw exceptions if the cause is a duplicate key.
Throwing an exception will cause the message to be retried. If the exception
is thrown continually, then the message will eventually end up in the dead
letter queue.

This works nicely most of the time - ie, if the DB is down for a while,
we won't be able to write to it, things will fail and the messages end up
in the dead letter queue, ready to be replayed at a later date.

However, if the message is a duplicate of an event which is already in the
database, then we don't want to keep retrying (as the event is already handled,
and retrying will definately not solve anything). These messages also should
not end up in the dead letter queue, as we want to reserve that queue for
when bad things happen that require attention.
If the event is already in the DB, then we should consider the event 'handled',
log a warning about this having happened, and move on to the next message.

Solo: @michaelwalker