ABD-42: Decrypt strings fetched from the queue.
Its a requirement to never have plaintext messages at rest in the queue.
As a result, we need to encrypt messages before writing to the queue, and
decrypt messages here before writing to the database (which is encrypted
by RDS).

We've ruled out using KMS for encryption/decryption of the messages directly,
as the latency of encryption requests is too high to be acceptable on the
hub side (sending to SQS needs to be a fast process).
Instead, we use KMS to encrypt a symmetric key, which we then store in
S3. We can retreive this key (using KMS again to decrypt) and then use
this key to manually encrypt or decrypt our messages.
This method dramatically reduces latency (as we only need to fetch from
KMS at startup) but does make it much harder to rotate keys.

For the cipher, we have chosen AES (in CBC mode). This is a fast, secure and widely
implemented cipher which will allow us to easily encrypt in the Java
event-emitter and decrypt in the python event-recorder.

Authors: @michaelwalker, @andy-c-jones