ABD-41: Provide linux binaries for lambda runtime.
Some of our python libraries are not pure python. Instead they rely on
some C binaries.

These C binaries are compiled for a specific OS (and version of python) and
are not valid in other environments.

Therefore binaries built on our dev machines will not work on Lambda.

As a workaround, we have created all the required binaries on a linux VM,
and have added them to source control. Our package task will use these
binaries in preference to any which are created on the host system.

This is very unpleasant, and we should attempt to find a less horrible
solution moving forward.

Authors: @michaelwalker @andy-c-jones