353: Store event's timestamp in datetime format
Currently, a database is rejecting event messages since their
timestamp's format is integer and the database expects it to be in date
time format. This change updates an event recorder system to convert the
event message's timestamp from integer into datetime so that the message
can then be stored in the database without errors.

Authors: @adityapahuja