TRELLO-awtbKHKO: Write billing events and fraud events to billing schema
PostgreSQL has a billing schema which is same as an billingDb database
in MongoDB. The schema has a billing_events table and a fraud_events
table. Both billing_events table and fraud_events table have same
attributes as a billingEvents collection and fraudEvents collection's
attributes from MongoDB respectively. This change enables an event
recorder service to write a billing event and a fraud event to the
billing_events table and the fraud_events table in the billing schema in
PostgreSQL respectively.

The event recorder service sends idp authentication succeeded events as
billing events to billing_events table and it sends fraud detected
events as fraud events to audit_events table.

There is billingIdps collection in billingDb and the collection is
excluded from this change since it contains a list of identity
providers' entity ids. Billing application receives a billing event or a
fraud event and checks its details' identity provider entity id is
present in the billingIdps collection. If the identity provider entity
id is not present, it adds to billingIdps collections. The event
recorder service does not need do this since identity provider entity id
is already present in billingEvents collection and fraudEvent
collection.

Authors: @adityapahuja