Switch to using Dockerfile for database container
The database schemas used for initialisation of the test image will
soon be moved to a new home. By switching to a Dockerfile, it is a
simple matter of cloning the new repo and copying the schema files
into the initialisation folder PostgreSQL uses during initial setup.

Note that the Dockerfile and changes to the docker-compose
configuration are taken from the `verify-billing-reporter` repo with
minimal changes.